create_trial_type_by_volume_list

create_trial_type_by_volume_list

create_trial_type_by_volume_list(
  events,
  tr,
  n_volumes,
  thr = 0.51,
  unlabeled_trial_type = "fixation"
)

Arguments

events
events with columns of onset, duration, and trial_type

tr
repetition time (tr) in seconds

n_volumes
number of volumes or time points

thr
threshold or percentage needed to label volume as trial_type

unlabeled_trial_type
trial_type for volumes without names (default: "fixation")

Value

data with columns of volume and trial_type

Examples

#' library(dplyr)

# this examples uses the following data from openneuro.org
# https://openneuro.org/datasets/ds000171/versions/00001

# in this dataset, participants listen to either positive music,
# negative music, or tones (control). Positive and negative musical stimulus
# are presented for 31.5 seconds and tones are presented for 33 seoncds.
# After each stimulus, participants have 3 seconds to report their valence
# (from very sad to very happy) and arousal (from very passive to very #
# active). the repitition time (tr) for this dataset was 3 and the number of
# volumes collected for this was run 105.

# mri parameters
tr <- 3
n_volumes <- 105

# read events from openneuro.org
psy_events <- readr::read_tsv(url("https://openneuro.org/crn/datasets/ds000171/snapshots/00001/files/sub-control01:func:sub-control01_task-music_run-1_events.tsv")) %>%
  mutate(trial_type = as.factor(trial_type))

#> Rows: 18 Columns: 3

#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr (1): trial_type
#> dbl (2): onset, duration

#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

# orthogonal contrast code that test:
# 1. stimulus vs response
# 2. music vs tones
# 3. positive music vs negative music
psy_contrast_table <- as.data.frame(cbind(stimulus_vs_response = c(1, 1, -3, 1)/4,
                                          music_vs_tones = c(1, 1, 0, -2)/3,
                                          positive_music_vs_negative_music = c(-1, 1, 0, 0)/2))

# for this example, we will choose SPM's default canonical hrf and upsample factor of 16
upsample_factor <- 16
hrf <- create_hrf_afni("spmg1", tr, upsample_factor)

# run function
psy_trial_type_by_volume <- create_trial_type_by_volume_list(events = psy_events,
                          tr = tr,
                          n_volumes = n_volumes,
                          unlabeled_trial_type = "response")

head(psy_trial_type_by_volume)

#> # A tibble: 6 × 2
#>   volume trial_type
#>    <int> <fct>     
#> 1      1 tones     
#> 2      2 tones     
#> 3      3 tones     
#> 4      4 tones     
#> 5      5 tones     
#> 6      6 tones     

summary(psy_trial_type_by_volume$trial_type)

#> negative_music positive_music       response          tones 
#>             20             20             10             55 

# references:
# Lepping RJ, Atchley RA, Chrysikou E, Martin LE, Clair AA, Ingram RE, et al. Neural processing of emotional musical and nonmusical stimuli in depression.  PlosONE.  In Press.
# Lepping RJ, Atchley RA, Savage CR. Development of a Validated Emotionally Provocative Musical Stimulus Set for Research. Psychology of Music. 2015; online before print. doi:10.1177/0305735615604509.
# Lepping RJ, Atchley RA, Martin LE, Patrician TM, Stroupe N, Brooks WM, et al. Limbic responses to positive and negative emotionally evocative music: An fMRI study.  Society for Neuroscience annual meeting 2012 Oct 13-18; New Orleans, LA 2012.
# Lepping RJ, Atchley RA, Martin LE, Brooks WM, Clair AA, Ingram RE, et al. Musical and nonmusical sounds evoke different patterns of neural activity: An fMRI study.  Psychonomic Society annual meeting; 2012 Nov 14-18; Minneapolis, MN 2012.
# Lepping RJ, Atchley RA, Martin LE, Patrician TM, Ingram RE, Clair AA, et al. The effect of musical experiences and musical training on neural responses to emotionally evocative music and non-musical sounds.  Social and Affective Neuroscience Society annual conference 2013 Apr 12-13; San Francisco, CA 2013
# Lepping RJ, Atchley RA, Patrician TM, Stroupe NN, Martin LE, Ingram RE, et al. Music to my ears: Neural responsiveness to emotional music and sounds in depression.  Society of Biological Psychiatry annual scientific convention 2013 May 16-18; San Francisco, CA 2013.