data_wrangling

uber data wrangling wrapper function to create data for each step of data wrangling for each psy, phys, and ppi variables to ultimately create the design matrix. This function is a wrapper of create_hrf_afni(), create_psy_var(), create_phys_var(), create_ppi_var(), and create_design_matrix()

data_wrangling(psy_events_data, psy_contrast_table, phys_data,
  detrend_factor, hrf_name, tr, n_volumes, upsample_factor = NULL,
  deconvolve = TRUE, nuisance_var = NULL)

Arguments

psy_events_data
events data with columns of onset, duration, and trial_type

psy_contrast_table
contrast code

phys_data
time series data of physiological variable (seed ROI)

detrend_factor
factor to detrend time series

hrf_name
name of hemodynamic response function (hrf) to use during convolution and deconvolution step

tr
repition time (tr) in seconds

n_volumes
number of volumes or time points

upsample_factor
factor to upsample psy and phys data for convolution and deconvolution step

deconvolve
perform deconvolution of phys and convolve of ppi term (also known as deconvolve-convolve or reconvolve step)

nuisance_var
data of nuisance variables to include in design matrix

Examples

# to be added